+++ src/lib.rs
@@ -69,7 +69,8 @@ pub use device::{Device, Queue};
 pub use instance::{Instance, VkError};
 pub use physical_device::{PhysicalDevice, PhysicalDeviceProperties, VideoExtensionSupport};
 pub use video::{
-    query_video_decode_av1_capabilities, query_video_decode_h264_capabilities,
+    av1_level_label, av1_profile_label, h264_level_label, h264_profile_label, h265_level_label,
+    h265_profile_label, query_video_decode_av1_capabilities, query_video_decode_h264_capabilities,
     query_video_decode_h265_capabilities, VideoDecodeAV1Capabilities, VideoDecodeH264Capabilities,
     VideoDecodeH265Capabilities, VideoSession,
 };